tracking.py
import cv2
import numpy as np
import logging

logger = logging.getLogger(__name__)

class VehicleTracker:
    def __init__(self, max_age=30, n_init=3):
        # Placeholder for DeepSort tracker
        self.tracked_vehicles = {}
        self.next_id = 0

    def update(self, detections, frame):
        """
        Update tracker with new detections (placeholder implementation)
        detections: list of dicts with 'bbox' and 'confidence'
        Returns: list of tracked objects with IDs
        """
        # Simple tracking placeholder - assign new IDs to detections
        tracked_objects = []
        for det in detections:
            bbox = det['bbox']
            self.next_id += 1
            tracked_objects.append({
                'id': self.next_id,
                'bbox': bbox,
                'centroid': ((bbox[0] + bbox[2]) // 2, (bbox[1] + bbox[3]) // 2),
                'original_detection': det
            })

        return tracked_objects

class RiderTracker:
    def __init__(self, max_age=30, n_init=3):
        # Placeholder for DeepSort tracker
        self.tracked_riders = {}
        self.next_id = 0

    def update(self, detections, frame):
        """
        Update tracker with rider detections (placeholder implementation)
        """
        tracked_riders = []
        for det in detections:
            bbox = det['bbox']
            self.next_id += 1
            tracked_riders.append({
                'id': self.next_id,
                'bbox': bbox,
                'centroid': ((bbox[0] + bbox[2]) // 2, (bbox[1] + bbox[3]) // 2),
                'original_detection': det
            })

        return tracked_riders
