requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
deep-sort-realtime==1.3.2
pytesseract==0.3.10
streamlit==1.28.1
pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
scipy==1.11.3
matplotlib==3.7.2
requests==2.31.0
